backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
httpx
python-dotenv
openai>=1.0.0
requests
duckduckgo-search
boto3
h2
pytest
pytest
Pillow
SQLAlchemy>=2.0,<3.0